requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# AI / LLM
google-genai==1.14.0
langgraph==0.4.1
langchain-core>=0.3.52,<0.4.0
langchain-google-genai==2.1.3

# Resume parsing
PyPDF2==3.0.1

# Frontend
streamlit==1.42.2
streamlit-webrtc==0.47.9
streamlit-ace==0.1.1
plotly==6.0.1
av==14.0.0

# Utilities
httpx==0.28.1
numpy==2.2.3
